% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.statlab_config}
\alias{print.statlab_config}
\title{Afficher un objet de configuration statlab}
\usage{
\method{print}{statlab_config}(x, ...)
}
\arguments{
\item{x}{Un objet \code{statlab_config}.}

\item{...}{Ignore.}
}
\value{
\code{x}, de maniere invisible.
}
\description{
Afficher un objet de configuration statlab
}
