% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{st_plot_bar}
\alias{st_plot_bar}
\title{Diagramme en barres d'une variable categorielle, avec effectifs et pourcentages}
\usage{
st_plot_bar(data, variable, group = NULL, config, test_result = NULL)
}
\arguments{
\item{data}{Un \code{data.frame}.}

\item{variable}{Nom de la variable categorielle (chaine).}

\item{group}{Optionnel, nom d'une variable de groupement (chaine).}

\item{config}{Un objet \code{statlab_config}.}

\item{test_result}{Optionnel, resultat de \code{\link[=st_compare]{st_compare()}}.}
}
\value{
Un objet \code{ggplot}.
}
\description{
Diagramme en barres d'une variable categorielle, avec effectifs et pourcentages
}
