% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{st_table1_csv}
\alias{st_table1_csv}
\title{Exporter le Tableau 1 en tableau brut}
\usage{
st_table1_csv(table1)
}
\arguments{
\item{table1}{L'objet retourne par \code{\link[=st_table1]{st_table1()}}.}
}
\value{
Un data.frame (une ligne par caracteristique/modalite affichee).
}
\description{
Exporter le Tableau 1 en tableau brut
}
