% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{st_table1_flextable}
\alias{st_table1_flextable}
\title{Convertir le Tableau 1 en flextable}
\usage{
st_table1_flextable(table1)
}
\arguments{
\item{table1}{L'objet retourne par \code{\link[=st_table1]{st_table1()}}.}
}
\value{
Un objet \code{flextable}.
}
\description{
Conversion NATIVE (jamais une image) pour insertion editable dans un
document Word.
}
