% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statlab-package.R
\docType{package}
\name{statlab-package}
\alias{statlab}
\alias{statlab-package}
\title{statlab: Outils Internes de Statistique et de Reporting}
\description{
Boite a outils interne fournissant des regles, gabarits et schemas
standardises pour les traitements statistiques et le reporting.
}
\details{
Ce package est actuellement un squelette : aucune fonctionnalite
metier n'est encore implementee. L'arborescence prevoit :

\itemize{
\item \code{inst/rules} : regles metier packagees avec le produit final.
\item \code{inst/templates} : gabarits de documents et de rapports.
\item \code{inst/schema} : schemas de validation des donnees.
}
}
\author{
\strong{Maintainer}: Dedji Abdoul Batinou \email{societedilari@gmail.com}

Authors:
\itemize{
  \item Dedji Abdoul Batinou \email{societedilari@gmail.com}
}

}
\keyword{internal}
